Docker: update nginx to fix mod_zip module compliation warning
Update nginx 1.9.9 to 1.14.0.

As figured out in #51 there are two warnings [1] regarding the mod_zip
module. Verified with older stable Nginx versions too like 1.8.1 and got
the same result.
Newer ones like 1.10.3 and 1.12.2 aren´t affected. As 1.9.9 is actually
pretty old I´ve bumped the version to the latest and greatest stable
version which is 1.14.0 at the time of writing.

[1]: https://travis-ci.org/openseedbox/openseedbox/jobs/369439939#L1544