Docker: merge OpenID Connect PR to work around Google Sign-in shutdown
While pull request #87 is not ready for merge, it works nicely.
To avoid the impact of Google Sign-In shutdown in March 31st, 2023
just merge the PR for Docker image until it's ready.

KEYCLOAK_CLIENTID and GITHUB_CLIENTID is needed to hide the unconfigured
auth providers.

To avoid "fatal: refusing to merge unrelated histories" errors,
let's unshallow some of the history with the date of the previous commit.

See #76